microservices/core-api/requirements.txt
# Core API Service Dependencies
# Minimal dependencies for authentication and API gateway

# Core FastAPI and web dependencies
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
httpx

# Database
asyncpg
psycopg2-binary
psycopg[binary]>=3.1.0

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# Environment and utilities
python-dotenv
